+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+
+## 0.2.4 (in development)
+- support for link tolls in DialB
+- Dial B reimplemented on the link-turn graph and destination based
+- heuristic for finding k-shortest path with limited overlap
+
+
 ## 0.2.3 (08.09.2022)
 - bugfix polygon creation 
 - adding SUE assignment